Backport to NW 7.0 + added export_to_table
I removed the method chaining in the unit tests and a few abap key words
that don't exist in NW 7.0.

I also added an export_to_table method. We provide a lot of web services
to a Java team and we often return the error messages to them to be
handled so having an easy way to get them as bapiret2_t is awesome :)