Create XPRA logger when called from XPRA report
XPRA reports are executed in client 000 in a job named RDDEXECL. 
A minimal amount of changes have been made so to return an XPRAlogger when the report is executed as an XPRA report.

There is much room for improvement, such as moving functionality to a common base class; adding (additional) messages with context information in an XPRA logger, etc etc.
But this covers our basic needs for the moment.